Move getCanonicalRoutes
